requirements.txt
aiogram>=3.22.0,<4.0
python-dotenv>=1.0,<2.0
pymongo>=4.8,<5.0
